+++ bin/miden-cli/src/config.rs
@@ -2,6 +2,7 @@ use core::fmt::Debug;
 use std::{
     fmt::Display,
     path::{Path, PathBuf},
+    str::FromStr,
 };
 
 use figment::{
@@ -9,6 +10,7 @@ use figment::{
     value::{Dict, Map},
 };
 use miden_client::rpc::Endpoint;
+use miden_objects::{NetworkIdError, account::NetworkId};
 use serde::{Deserialize, Serialize};
 
 const TOKEN_SYMBOL_MAP_FILEPATH: &str = "token_symbol_map.toml";
@@ -33,6 +35,8 @@ pub struct CliConfig {
     pub remote_prover_endpoint: Option<CliEndpoint>,
     /// Path to the directory from where account component template files will be loaded.
     pub component_template_directory: PathBuf,
+    /// Network for use in bech32 account ID enconding.
+    pub network: Network,
 }
 
 // Make `ClientConfig` a provider itself for composability.
@@ -67,6 +71,7 @@ impl Default for CliConfig {
             token_symbol_map_filepath: Path::new(TOKEN_SYMBOL_MAP_FILEPATH).to_path_buf(),
             remote_prover_endpoint: None,
             component_template_directory: Path::new(DEFAULT_COMPONENT_TEMPLATE_DIR).to_path_buf(),
+            network: Network::Localhost,
         }
     }
 }
@@ -147,5 +152,57 @@ impl<'de> Deserialize<'de> for CliEndpoint {
     {
         let endpoint = String::deserialize(deserializer)?;
         CliEndpoint::try_from(endpoint.as_str()).map_err(serde::de::Error::custom)
+    }
+}
+
+// NETWORK
+// ================================================================================================
+
+/// Represents the network to which the client connects. It is used to determine the RPC endpoint
+/// and network ID for the CLI.
+#[derive(Debug, Clone, Deserialize, Serialize)]
+pub enum Network {
+    Custom(String),
+    Devnet,
+    Localhost,
+    Testnet,
+}
+
+impl FromStr for Network {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_str() {
+            "devnet" => Ok(Network::Devnet),
+            "localhost" => Ok(Network::Localhost),
+            "testnet" => Ok(Network::Testnet),
+            custom => Ok(Network::Custom(custom.to_string())),
+        }
+    }
+}
+
+impl Network {
+    /// Converts the Network variant to its corresponding RPC endpoint string
+    #[allow(dead_code)]
+    pub fn to_rpc_endpoint(&self) -> String {
+        match self {
+            Network::Custom(custom) => custom.clone(),
+            Network::Devnet => Endpoint::devnet().to_string(),
+            Network::Localhost => Endpoint::default().to_string(),
+            Network::Testnet => Endpoint::testnet().to_string(),
+        }
+    }
+
+    /// Converts the Network variant to its corresponding [`NetworkId`]
+    #[allow(dead_code)]
+    pub fn to_network_id(&self) -> Result<NetworkId, NetworkIdError> {
+        let network_id = match self {
+            Network::Custom(_) => NetworkId::new("mcst")?,
+            Network::Devnet => NetworkId::Devnet,
+            Network::Localhost => NetworkId::new("mlcl")?,
+            Network::Testnet => NetworkId::Testnet,
+        };
+
+        Ok(network_id)
     }
 }